Fixed warning.
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@11501 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/kernel/libroot/os/debug.c
@@ -91,7 +91,7 @@ get_why_stopped_string(debug_why_stopped whyStopped, char *buffer,
 	if (!buffer || bufferSize <= 0)
 		return;
 
-	if (whyStopped >= 0 && whyStopped < sDebugWhyStringCount)
+	if ((int32)whyStopped >= 0 && (int32)whyStopped < sDebugWhyStringCount)
 		strlcpy(buffer, sDebugWhyStrings[whyStopped], bufferSize);
 	else
 		snprintf(buffer, bufferSize, "Unknown reason %ld", (int32)whyStopped);
